services/dashboard-usine/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
httpx==0.25.2
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
kafka-python==2.0.2
reportlab==4.0.7
pandas==2.1.3
numpy==1.26.2
geopandas==0.14.1
shapely==2.0.2
geoalchemy2==0.14.2

